requirements.txt
Flask
flask-socketio
Flask-SQLAlchemy
flask-bcrypt
Flask-Limiter
Flask-CORS
python-dotenv
pyotp
pytz
requests
pandas
SQLAlchemy
cachetools
gunicorn
psycopg2-binary

numpy>=1.22.2 # not directly required, pinned by Snyk to avoid a vulnerability
werkzeug>=2.3.8 # not directly required, pinned by Snyk to avoid a vulnerability
